+++ NativeAPI/MessageObject.cpp
@@ -4,5 +4,6 @@
 struct MessageObject {
 	std::string body;
 	uint64_t messageTag;
+	int priority;
 	std::map<int, std::string> msgProps;
 };