+++ NativeAPI/RabbitMQClient.cpp
@@ -124,7 +124,7 @@ bool RabbitMQClient::deleteExchange(const std::string& name, bool ifunused) {
 	return result;
 }
 
-std::string RabbitMQClient::declareQueue(const std::string& name, bool onlyCheckIfExists, bool durable, bool autodelete) {
+std::string RabbitMQClient::declareQueue(const std::string& name, bool onlyCheckIfExists, bool durable, bool autodelete, uint16_t maxPriority) {
 
 	updateLastError("");
 
@@ -133,7 +133,12 @@ std::string RabbitMQClient::declareQueue(const std::string& name, bool onlyCheck
 		return "";
 	}
 
-	channelLoc->declareQueue(name, (onlyCheckIfExists ? AMQP::passive : 0) | (durable ? AMQP::durable : 0) | (durable ? AMQP::durable : 0) | (autodelete ? AMQP::autodelete : 0))
+	AMQP::Table args = AMQP::Table();
+	if (maxPriority != 0) {
+		args.set("x-max-priority", maxPriority);
+	}
+
+	channel->declareQueue(name, (onlyCheckIfExists ? AMQP::passive : 0) | (durable ? AMQP::durable : 0) | (durable ? AMQP::durable : 0) | (autodelete ? AMQP::autodelete : 0), args)
 		.onSuccess([this]()
 	{
 		handler->quit();
@@ -266,6 +271,7 @@ bool RabbitMQClient::basicPublish(std::string& exchange, std::string& routingKey
 		if (!msgProps[CLUSTER_ID].empty()) envelope.setClusterID(msgProps[CLUSTER_ID]);
 		if (!msgProps[EXPIRATION].empty()) envelope.setExpiration(msgProps[EXPIRATION]);
 		if (!msgProps[REPLY_TO].empty()) envelope.setReplyTo(msgProps[REPLY_TO]);
+		if (priority != 0) envelope.setPriority(priority);
 
 		publChannel.publish(exchange, routingKey, envelope);
 		handler->quit();
@@ -317,6 +323,7 @@ std::string RabbitMQClient::basicConsume(const std::string& queue, const int _se
 		msgOb->msgProps[EXPIRATION] = message.expiration();
 		msgOb->msgProps[REPLY_TO] = message.replyTo();
 		msgOb->messageTag = deliveryTag;
+		msgOb->priority = message.priority();
 
 		readQueue->push(msgOb);
 	})
@@ -353,6 +360,7 @@ bool RabbitMQClient::basicConsumeMessage(std::string& outdata, uint16_t timeout)
 			msgProps[CLUSTER_ID] = read->msgProps[CLUSTER_ID];
 			msgProps[EXPIRATION] = read->msgProps[EXPIRATION];
 			msgProps[REPLY_TO] = read->msgProps[REPLY_TO];
+			priority = read->priority;
 
 			confirmQueue->push(read);
 			return true;
@@ -430,10 +438,19 @@ bool RabbitMQClient::basicCancel() {
 	return true;
 }
 
-WCHAR_T* RabbitMQClient::getLastError() noexcept {
-	return LAST_ERROR;
+bool RabbitMQClient::setPriority(int _priority) {
+	priority = _priority;
+	return true;
 }
 
+int RabbitMQClient::getPriority() {
+	return priority;
+}
+
+WCHAR_T* RabbitMQClient::getLastError() noexcept
+{
+	return LAST_ERROR;
+}
 
 void RabbitMQClient::updateLastError(const char* text) {
 	LAST_ERROR = new wchar_t[strlen(text) + 1];